requirements.txt
aiohttp
aiosignal
altair
aniso8601
async-timeout
attrs
bitarray
blinker
blis
cachetools
catalogue
certifi
cffi
chardet
charset-normalizer
click
colorama
cryptography
cymem
decorator
docx2txt
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.0.0/en_core_web_sm-3.0.0-py3-none-any.whl
Flask
Flask-Cors
Flask-RESTful
frozenlist
gitdb
GitPython
idna
importlib-metadata
itsdangerous
Jinja2
joblib
jsonschema
langcodes
markdown-it-py
MarkupSafe
mdurl
multidict
murmurhash
mysql-connector
nltk
numpy
openai
packaging
pafy
pandas
pathy
pdfminer.six
pdfminer3
pdfplumber
pdfreader
phonenumbers
Pillow
plac
plotly
preshed
protobuf
pyarrow
pycosat
pycparser
pycryptodome
pydantic
pydeck
Pygments
Pympler
PyMySQL
pyresparser
pyrsistent
python-dateutil
python-dotenv
pytz
pytz-deprecation-shim
PyYAML
regex
requests
rich
six
smart-open
smmap
sortedcontainers
spacy
spacy-legacy
spacy-loggers
srsly
stemming
streamlit
streamlit-tags
tenacity
thinc
tika
toml
toolz
tornado
tqdm
typer
typing_extensions
tzdata
tzlocal
urllib3
validators
Wand
wasabi
watchdog
Werkzeug
yarl
youtube-dl
zipp